Specifying Scipy version
The setup script first installs numpy==1.15.4 from requirements.txt. While installing scipy, since the version is not mentioned, it takes the recent scipy version (1.4.1) which has dependency of numpy==1.18.1. This creates 2 different site-packages for numpy in the environment.
While running the  unit_tests.sh, it throws error for having two numpy versions in the same environment.
The existing version of scipy when setup.sh file was committed was scipy==1.3.1.
This commit will prevent the script from breaking.